handlers 컨벤션에 맞추어 validation depth줄이기 및 리팩토링
