+++ README.md
@@ -39,7 +39,9 @@
 - 리팩토링
   - 자바스크립트 컨벤션에 맞추어 validation depth줄이기 및 리팩토링 v
   - 폴더 재구성 v
-  - html tag 검토 (인풋 밸류, 글자 크기)
+  - handlers 컨벤션에 맞추어 validation depth줄이기 및 리팩토링 v
+  - html tag 검토 (인풋 밸류)
+  - css 검토
 
 ## 🤦🏻‍♀️ 고민한 지점
