+++ src/managers/handlers/mapPrintHandler.js
