+++ src/managers/render.js
